requirements.txt
streamlit>=1.32.0
bcrypt>=4.0.0
cryptography>=42.0.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
rapidfuzz>=3.0.0
sentence-transformers>=2.7.0
openpyxl>=3.1.0
reportlab>=4.0.0
plotly>=5.18.0
pandas>=2.0.0
anthropic>=0.25.0
